+++ data/src/main/kotlin/com/wepli/data/chart/datasource/remote/ChartSupabaseDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.wepli.data.chart.datasource.remote
 
 import com.wepli.data.chart.response.ChartListResponse
 import com.wepli.core.kotlin.flow.FlowResult
-import com.wepli.data.SupabaseTable
+import com.wepli.data.supabase.SupabaseConstants
 import io.github.jan.supabase.SupabaseClient
 import io.github.jan.supabase.postgrest.postgrest
 import kotlinx.coroutines.flow.flow
@@ -14,7 +14,7 @@ class ChartSupabaseDataSourceImpl @Inject constructor(
 
     override suspend fun getTopChart(): FlowResult<ChartListResponse> = flow {
         val result = runCatching {
-            supabase.postgrest[SupabaseTable.CHART_TABLE]
+            supabase.postgrest[SupabaseConstants.CHART_TABLE]
                 .select()
                 .decodeList<ChartListResponse.ChartResponse>()
                 .run(::ChartListResponse) // 기존 API 형식과 동일하게 ChartListResponse로 래핑해서 반환
