+++ data/src/main/kotlin/com/wepli/data/user/datasource/UserSupabaseDataSourceImpl.kt
@@ -1,12 +1,11 @@
 package com.wepli.data.user.datasource
 
 import com.wepli.core.kotlin.flow.FlowResult
-import com.wepli.data.SupabaseTable
+import com.wepli.data.supabase.SupabaseConstants
 import com.wepli.data.user.response.UserResponse
 import io.github.jan.supabase.SupabaseClient
 import io.github.jan.supabase.postgrest.postgrest
 import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.flowOf
 import model.user.User
 import javax.inject.Inject
 
@@ -16,7 +15,7 @@ class UserSupabaseDataSourceImpl @Inject constructor(
 
     override suspend fun getUserById(id: String): FlowResult<UserResponse> = flow {
         val result = runCatching {
-            supabase.postgrest[SupabaseTable.USER_TABLE]
+            supabase.postgrest[SupabaseConstants.USER_TABLE]
                 .select {
                     filter {
                         eq("id", id)
@@ -30,7 +29,7 @@ class UserSupabaseDataSourceImpl @Inject constructor(
 
     override suspend fun updateUser(user: User): FlowResult<Unit> = flow {
         val result = runCatching {
-            supabase.postgrest[SupabaseTable.USER_TABLE]
+            supabase.postgrest[SupabaseConstants.USER_TABLE]
                 .update(
                     {
                         set("username", user.nickname)
